Organize ChatPanel tests
