+++ apps/web/package.json
@@ -6,7 +6,7 @@
     "dev": "vite",
     "build": "tsc -b && vite build",
     "preview": "vite preview",
-    "test": "vitest run src/appData/domain/index.test.ts src/appData/session/useWorkspaceSession.test.tsx src/appData/sync/syncLifecycleObservation.test.ts src/localDb/core.migrations.test.ts src/localDb/reviews.test.ts src/localDb/progress.test.ts src/localDb/reviewSchedule.test.ts src/localDb/cards.test.ts src/localDb/feedback.test.ts src/feedback/**/*.test.ts src/screens/review/**/*.test.ts src/screens/review/**/*.test.tsx src/screens/settings/**/*.test.tsx src/screens/progress/ProgressScreen.render.test.tsx src/appData/progress/progressSource.summarySeries.test.tsx src/appData/progress/progressSource.cache.test.tsx src/appData/progress/progressSource.reviewSchedule.test.tsx src/appData/progress/progressSource.lifecycle.test.tsx src/appData/progress/progressInvalidation.test.tsx src/api/endpoints.test.ts src/api/transport.test.ts src/chat/streaming/liveStream.test.ts src/chat/attachments/*.test.ts src/chat/ChatPanel/*.test.tsx src/chat/composer/*.test.ts src/accountDeletion.test.ts src/observability/instrument.test.ts",
+    "test": "vitest run src/appData/domain/index.test.ts src/appData/session/useWorkspaceSession.test.tsx src/appData/sync/syncLifecycleObservation.test.ts src/localDb/core.migrations.test.ts src/localDb/reviews.test.ts src/localDb/progress.test.ts src/localDb/reviewSchedule.test.ts src/localDb/cards.test.ts src/localDb/feedback.test.ts src/feedback/**/*.test.ts src/screens/review/**/*.test.ts src/screens/review/**/*.test.tsx src/screens/settings/**/*.test.tsx src/screens/progress/ProgressScreen.render.test.tsx src/appData/progress/progressSource.summarySeries.test.tsx src/appData/progress/progressSource.cache.test.tsx src/appData/progress/progressSource.reviewSchedule.test.tsx src/appData/progress/progressSource.lifecycle.test.tsx src/appData/progress/progressInvalidation.test.tsx src/api/endpoints.test.ts src/api/transport.test.ts src/chat/streaming/liveStream.test.ts src/chat/attachments/*.test.ts src/chat/ChatPanel/tests/*.test.tsx src/chat/composer/*.test.ts src/accountDeletion.test.ts src/observability/instrument.test.ts",
     "test:e2e": "FLASHCARDS_E2E_TARGET=prod playwright test",
     "test:e2e:local": "node ./scripts/check-local-e2e-stack.mjs && FLASHCARDS_E2E_TARGET=local playwright test",
     "test:e2e:prod": "FLASHCARDS_E2E_TARGET=prod playwright test"
