+++ apps/web/src/chat/ChatPanel/tests/ChatPanel.send-error-handling.test.tsx
@@ -7,7 +7,7 @@ import {
   readStoredDraftInputText,
   setupChatPanelTest,
   startChatRunMock,
-} from "./ChatPanelTestSupport";
+} from "./support/ChatPanelTestSupport";
 
 const {
   flushAsync,
