+++ apps/web/src/chat/ChatPanel/tests/ChatPanel.send-session-recovery.test.tsx
@@ -9,7 +9,7 @@ import {
   readStoredDraftInputText,
   setupChatPanelTest,
   startChatRunMock,
-} from "./ChatPanelTestSupport";
+} from "./support/ChatPanelTestSupport";
 
 const {
   flushAsync,
