+++ apps/web/src/chat/ChatPanel/tests/ChatPanel.send-terminal-reconcile.test.tsx
@@ -11,7 +11,7 @@ import {
   queryChatStopButton,
   setupChatPanelTest,
   startChatRunMock,
-} from "./ChatPanelTestSupport";
+} from "./support/ChatPanelTestSupport";
 
 const {
   flushAsync,
